+++ apps/web/src/_pages/moim-edit/actions.ts
@@ -3,7 +3,7 @@
 import { isRedirectError } from "next/dist/client/components/redirect-error";
 import { redirect } from "next/navigation";
 import { updateMoim } from "@/_pages/moim-edit/use-cases/moim-update";
-import { parseMoimFormData } from "@/features/moim-create";
+import { parseMoimFormData } from "@/features/moim-create/model/parse-moim-form-data";
 import type { MoimCreateFormValues } from "@/features/moim-create/model/schema";
 import { isAuth } from "@/shared/api/server";
 import { ROUTES } from "@/shared/config/routes";
