+++ apps/web/src/features/moim-create/index.ts
@@ -1,3 +1,2 @@
-export { parseMoimFormData } from "./model/parse-moim-form-data";
 export { MoimCreateForm } from "./ui/moim-create-form";
 export { MoimFormFields } from "./ui/moim-form-fields";
